+++ src/cmd/install_fish_autocompletion.go
@@ -76,6 +76,8 @@ type autocompleteDefinition struct {
 
 func buildAutocompletionDefinition() string {
 	commands := []autocompleteDefinition{
+		{name: "abort", description: abortCmd.Short},
+		{name: "continue", description: configCommand.Short},
 		{name: "hack", description: hackCmd.Short},
 		{name: "kill", description: killCommand.Short},
 		{name: "new-pull-request", description: newPullRequestCommand.Short},
@@ -84,11 +86,7 @@ func buildAutocompletionDefinition() string {
 		{name: "repo", description: repoCommand.Short},
 		{name: "ship", description: shipCmd.Short},
 		{name: "sync", description: syncCmd.Short},
-	}
-	options := []autocompleteDefinition{
-		{name: "abort", description: abortFlagDescription},
-		{name: "continue", description: continueFlagDescription},
-		{name: "undo", description: undoFlagDescription},
+		{name: "undo", description: undoCmd.Short},
 	}
 
 	commandsSpaceSeparated := ""
@@ -99,12 +97,8 @@ func buildAutocompletionDefinition() string {
 	for _, command := range commands {
 		commandAutocompletion += fmt.Sprintf("complete --command git --arguments '%s' --description '%s' --condition '__fish_complete_git_town_no_command' --no-files\n", command.name, command.description)
 	}
-	optionAutocompletion := ""
-	for _, option := range options {
-		optionAutocompletion += fmt.Sprintf("complete --command git --long-option '%s' --description '%s' --no-files\n", option.name, option.description)
-	}
 
-	return fmt.Sprintf(fishAutocompletionTemplate, commandsSpaceSeparated, commandAutocompletion, optionAutocompletion)
+	return fmt.Sprintf(fishAutocompletionTemplate, commandsSpaceSeparated, commandAutocompletion)
 }
 
 func init() {