+++ src/cmd/undo.go
@@ -1,0 +1,33 @@
+package cmd
+
+import (
+	"github.com/Originate/git-town/src/git"
+	"github.com/Originate/git-town/src/steps"
+	"github.com/Originate/git-town/src/util"
+
+	"github.com/spf13/cobra"
+)
+
+var undoCmd = &cobra.Command{
+	Use:   "undo",
+	Short: "Undoes the last run git-town command",
+	Run: func(cmd *cobra.Command, args []string) {
+		runState := steps.LoadPreviousRunState()
+		if runState == nil || runState.IsUnfinished() {
+			util.ExitWithErrorMessage("Nothing to undo")
+		}
+		undoRunState := runState.CreateUndoRunState()
+		steps.Run(&undoRunState)
+	},
+	Args: cobra.NoArgs,
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		return util.FirstError(
+			git.ValidateIsRepository,
+			validateIsConfigured,
+		)
+	},
+}
+
+func init() {
+	RootCmd.AddCommand(undoCmd)
+}