+++ src/git/printable.go
@@ -8,11 +8,13 @@ import (
 	"github.com/Originate/git-town/src/util"
 )
 
+var noneString = "[none]"
+
 // GetPrintableMainBranch returns a user printable main branch
 func GetPrintableMainBranch() string {
 	output := GetMainBranch()
 	if output == "" {
-		return "[none]"
+		return noneString
 	}
 	return output
 }
@@ -21,9 +23,21 @@ func GetPrintableMainBranch() string {
 func GetPrintablePerennialBranches() string {
 	output := strings.Join(GetPerennialBranches(), "\n")
 	if output == "" {
-		return "[none]"
+		return noneString
 	}
 	return output
+}
+
+// GetPrintablePerennialBranchTrees returns a user printable list of perennial branches trees
+func GetPrintablePerennialBranchTrees() string {
+	trees := []string{}
+	for _, perennialBranch := range GetPerennialBranches() {
+		trees = append(trees, GetPrintableBranchTree(perennialBranch))
+	}
+	if len(trees) == 0 {
+		return noneString
+	}
+	return strings.Join(trees, "\n")
 }
 
 // GetPrintableNewBranchPushFlag returns a user printable new branch push flag