+++ RELEASE_NOTES.md
@@ -2,6 +2,37 @@
 
 ## Unreleased
 
+#### New Features
+
+* `git town config`: print perennial branch trees
+
+## 7.1.1 (2018-04-09)
+
+#### Bug Fixes
+
+* strip colors from the output of git commands run internally. This caused errors if you had git configured with `color.ui=always`
+
+## 7.1.0 (2018-04-05)
+
+#### New Features
+
+* automatically remove outdated configuration
+
+## 7.0.0 (2018-04-03)
+
+#### BREAKING CHANGES
+
+* `git town config`: `reset` and `setup` are now subcommands instead of flags
+* `--abort`, `--continue`, `--skip`, `--undo` flags removed. Instead there are now top level commands `git town abort`, `git town continue`, `git town skip`, `git town undo`
+
+#### New Features
+
+* now catches when there is an unfinished state from a git town command that hit conflicts. If you try to run another git town command, you will be prompted on how to resolve the unfinished state. The unfinished state can be discarded and there is also a new top level command `git town discard` to delete the state of the last run command.
+
+#### Bug Fixes
+
+* skip perennial branch prompt if there are no options
+
 ## 6.0.2 (2018-01-26)
 
 #### Bug Fixes
@@ -314,7 +345,7 @@
 * improved error messages when run outside a git repository
 * improved setup wizard for initial configuration in a git repository
 * added [contribution guide](/CONTRIBUTING.md)
-* added [tutorial](/documentation/tutorial.md)
+* added tutorial
 
 ## 0.5.0 (2015-01-08)